19: rename negamax -> solver
