+++ Makefile
@@ -15,7 +15,7 @@ C_FLAGS = -g -std=c++11 -Wall
 OBJS = $(OBJ_DIR)/main.o \
       $(OBJ_DIR)/state.o \
       $(OBJ_DIR)/renderer.o \
-      $(OBJ_DIR)/negamax.o
+      $(OBJ_DIR)/solver.o
 
 
 .PHONY: all directories test clean
@@ -40,8 +40,8 @@ $(OBJ_DIR)/state.o: $(SRC_DIR)/state.cpp
 $(OBJ_DIR)/renderer.o: $(SRC_DIR)/renderer.cpp
 	$(CC) $(C_FLAGS) $(SRC_DIR)/renderer.cpp -c -o $(OBJ_DIR)/renderer.o
 
-$(OBJ_DIR)/negamax.o: $(SRC_DIR)/negamax.cpp
-	$(CC) $(C_FLAGS) $(SRC_DIR)/negamax.cpp -c -o $(OBJ_DIR)/negamax.o
+$(OBJ_DIR)/solver.o: $(SRC_DIR)/solver.cpp
+	$(CC) $(C_FLAGS) $(SRC_DIR)/solver.cpp -c -o $(OBJ_DIR)/solver.o
 
 $(OBJ_DIR)/test_main.o: $(TEST_DIR)/test_main.cpp
 	$(CC) $(C_FLAGS) $(TEST_DIR)/test_main.cpp -c -o $(OBJ_DIR)/test_main.o
@@ -49,12 +49,12 @@ $(OBJ_DIR)/test_main.o: $(TEST_DIR)/test_main.cpp
 $(TEST_DIR)/state: $(OBJ_DIR)/test_main.o $(TEST_DIR)/state.cpp $(OBJ_DIR)/state.o
 	$(CC) $(C_FLAGS) $(OBJ_DIR)/test_main.o $(TEST_DIR)/state.cpp $(OBJ_DIR)/state.o -o $(TEST_DIR)/state
 
-$(TEST_DIR)/negamax: $(OBJ_DIR)/test_main.o $(TEST_DIR)/negamax.cpp $(OBJ_DIR)/negamax.o $(OBJ_DIR)/state.o 
-	$(CC) $(C_FLAGS) $(OBJ_DIR)/test_main.o $(TEST_DIR)/negamax.cpp $(OBJ_DIR)/negamax.o $(OBJ_DIR)/state.o -o $(TEST_DIR)/negamax
+$(TEST_DIR)/solver: $(OBJ_DIR)/test_main.o $(TEST_DIR)/solver.cpp $(OBJ_DIR)/solver.o $(OBJ_DIR)/state.o 
+	$(CC) $(C_FLAGS) $(OBJ_DIR)/test_main.o $(TEST_DIR)/solver.cpp $(OBJ_DIR)/solver.o $(OBJ_DIR)/state.o -o $(TEST_DIR)/solver
 
-test: directories $(TEST_DIR)/state $(TEST_DIR)/negamax
+test: directories $(TEST_DIR)/state $(TEST_DIR)/solver
 	./$(TEST_DIR)/state
-	./$(TEST_DIR)/negamax
+	./$(TEST_DIR)/solver
 
 # Remove the executable
 clean:
